pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amplifier-module-tool-github"
version = "1.5.0"
description = "GitHub integration tool module for Amplifier"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Microsoft Corporation" }
]
keywords = ["amplifier", "llm", "agent", "github", "issues", "pull-requests", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

dependencies = [
    "amplifier-core",
    "PyGithub>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.entry-points."amplifier.tools"]
tool-github = "amplifier_module_tool_github:mount"

[project.urls]
Homepage = "https://github.com/microsoft/amplifier"
Documentation = "https://github.com/microsoft/amplifier"
Repository = "https://github.com/microsoft/amplifier-module-tool-github"
Issues = "https://github.com/microsoft/amplifier-module-tool-github/issues"

[tool.uv]
package = true

[tool.uv.sources]
amplifier-core = { git = "https://github.com/microsoft/amplifier-core", branch = "main" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["amplifier_module_tool_github"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["amplifier_module_tool_github"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
